@@ -1,168 +1,60 @@
 {
-  "name": "angular-starter",
-  "version": "6.0.0",
-  "description": "An Angular Webpack Starter kit featuring Angular (Router, Http, Forms, Services, Tests, E2E, Coverage), Karma, Protractor, Jasmine, Istanbul, TypeScript, and Webpack by AngularClass",
-  "keywords": [
-    "angular",
-    "angular2",
-    "angular4",
-    "webpack",
-    "typescript"
-  ],
-  "author": "Patrick Stapleton <patrick@angularclass.com>",
-  "homepage": "https://github.com/AngularClass/angular-starter",
+  "name": "coreui-angular",
+  "version": "1.0.0-alpha.6",
+  "description": "Open Source Bootstrap Admin Template",
+  "author": "",
+  "url": "http://coreui.io",
+  "copyright": "Copyright 2017 creativeLabs Łukasz Holeczek",
   "license": "MIT",
   "scripts": {
-    "build:aot:prod": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
-    "build:aot": "npm run build:aot:prod",
-    "build:dev": "npm run clean:dist && npm run webpack -- --config config/webpack.dev.js --progress --profile",
-    "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
-    "build:prod": "npm run clean:dist && npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
-    "build": "npm run build:dev",
-    "ci:aot": "npm run lint && npm run test && npm run build:aot && npm run e2e",
-    "ci:jit": "npm run lint && npm run test && npm run build:prod && npm run e2e",
-    "ci:nobuild": "npm run lint && npm test && npm run e2e",
-    "ci:testall": "npm run lint && npm run test && npm run build:prod && npm run e2e && npm run build:aot && npm run e2e",
-    "ci:travis": "npm run lint && npm run test && npm run build:aot && npm run e2e:travis",
-    "ci": "npm run ci:testall",
-    "clean:dll": "npm run rimraf -- dll",
-    "clean:aot": "npm run rimraf -- compiled",
-    "clean:dist": "npm run rimraf -- dist",
-    "clean:install": "npm set progress=false && npm install",
-    "clean": "npm cache clean --force && npm run rimraf -- node_modules doc coverage dist compiled dll",
-    "docker": "docker",
-    "docs": "npm run typedoc -- --options typedoc.json --exclude '**/*.spec.ts' ./src/",
-    "e2e:live": "npm-run-all -p -r server:prod:ci protractor:live",
-    "e2e:travis": "npm-run-all -p -r server:prod:ci protractor:delay",
-    "e2e": "npm-run-all -p -r server:prod:ci protractor",
-    "github-deploy:dev": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubDev",
-    "github-deploy:prod": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubProd",
-    "github-deploy": "npm run github-deploy:dev",
-    "lint": "npm run tslint \"src/**/*.ts\"",
-    "node": "node",
-    "postinstall": "npm run webdriver:update",
-    "postversion": "git push && git push --tags",
-    "preclean:install": "npm run clean",
-    "preversion": "npm test",
-    "protractor": "protractor",
-    "protractor:delay": "sleep 3 && npm run protractor",
-    "protractor:live": "protractor --elementExplorer",
-    "rimraf": "rimraf",
-    "server:dev:hmr": "npm run server:dev -- --inline --hot",
-    "server:dev": "npm run webpack-dev-server -- --config config/webpack.dev.js --open --progress --profile --watch --content-base src/",
-    "server:prod": "http-server dist -c-1 --cors",
-    "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
-    "server": "npm run server:dev",
-    "start:hmr": "npm run server:dev:hmr",
-    "start": "npm run server:dev",
-    "test": "npm run lint && karma start",
-    "tslint": "tslint",
-    "typedoc": "typedoc",
-    "version": "npm run build",
-    "watch:dev:hmr": "npm run watch:dev -- --hot",
-    "watch:dev": "npm run build:dev -- --watch",
-    "watch:prod": "npm run build:prod -- --watch",
-    "watch:test": "npm run test -- --auto-watch --no-single-run",
-    "watch": "npm run watch:dev",
-    "webdriver-manager": "webdriver-manager",
-    "webdriver:start": "npm run webdriver-manager start",
-    "webdriver:update": "webdriver-manager update",
-    "webpack-dev-server": "node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-    "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "~4.3.1",
-    "@angular/common": "~4.3.1",
-    "@angular/compiler": "~4.3.1",
-    "@angular/core": "~4.3.1",
-    "@angular/forms": "~4.3.1",
-    "@angular/http": "~4.3.1",
-    "@angular/platform-browser": "~4.3.1",
-    "@angular/platform-browser-dynamic": "~4.3.1",
-    "@angular/platform-server": "~4.3.1",
-    "@angular/router": "~4.3.1",
-    "@angularclass/hmr": "~1.2.2",
-    "@angularclass/hmr-loader": "^3.0.4",
-    "core-js": "^2.4.1",
-    "http-server": "^0.9.0",
-    "ie-shim": "^0.1.0",
-    "reflect-metadata": "^0.1.10",
-    "rxjs": "~5.0.2",
-    "zone.js": "0.8.14"
+    "@angular/common": "^4.3.6",
+    "@angular/compiler": "4.3.6",
+    "@angular/core": "4.3.6",
+    "@angular/forms": "4.3.6",
+    "@angular/http": "4.3.6",
+    "@angular/platform-browser": "4.3.6",
+    "@angular/platform-browser-dynamic": "4.3.6",
+    "@angular/router": "4.3.6",
+    "@angular/upgrade": "4.3.6",
+    "chart.js": "2.6.0",
+    "core-js": "2.5.0",
+    "moment": "2.18.1",
+    "ng2-charts": "1.6.0",
+    "ngx-bootstrap": "1.9.1",
+    "rxjs": "5.4.3",
+    "ts-helpers": "1.1.2",
+    "zone.js": "0.8.17"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~4.3.1",
-    "@types/hammerjs": "^2.0.34",
-    "@types/jasmine": "2.5.45",
-    "@types/node": "^7.0.39",
-    "@types/source-map": "^0.5.0",
-    "@types/uglify-js": "^2.6.28",
-    "@types/webpack": "^2.2.16",
-    "add-asset-html-webpack-plugin": "^1.0.2",
-    "angular2-template-loader": "^0.6.2",
-    "assets-webpack-plugin": "^3.5.1",
-    "awesome-typescript-loader": "~3.1.2",
-    "codelyzer": "~2.1.1",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.0.0",
-    "css-loader": "^0.28.0",
-    "exports-loader": "^0.6.4",
-    "expose-loader": "^0.7.3",
-    "extract-text-webpack-plugin": "~2.1.0",
-    "file-loader": "^0.11.1",
-    "find-root": "^1.0.0",
-    "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "^2.28.0",
-    "imports-loader": "^0.7.1",
-    "inline-manifest-webpack-plugin": "^3.0.1",
-    "istanbul-instrumenter-loader": "2.0.0",
-    "jasmine-core": "^2.5.2",
-    "karma": "^1.6.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-mocha-reporter": "^2.2.3",
-    "karma-remap-coverage": "^0.1.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
-    "ng-router-loader": "^2.1.0",
-    "ngc-webpack": "^3.2.0",
-    "node-sass": "^4.5.2",
-    "npm-run-all": "^4.0.2",
-    "optimize-js-plugin": "0.0.4",
-    "parse5": "^3.0.2",
-    "preload-webpack-plugin": "^1.2.2",
-    "protractor": "^5.1.1",
-    "raw-loader": "0.5.1",
-    "rimraf": "~2.6.1",
-    "sass-loader": "^6.0.3",
-    "script-ext-html-webpack-plugin": "^1.8.5",
-    "source-map-loader": "^0.2.1",
-    "string-replace-loader": "~1.2.0",
-    "style-loader": "^0.18.1",
-    "to-string-loader": "^1.1.5",
-    "ts-node": "^3.3.0",
-    "tslib": "^1.7.1",
-    "tslint": "~4.5.1",
-    "tslint-loader": "^3.5.2",
-    "typedoc": "^0.7.1",
-    "typescript": "~2.2.2",
-    "url-loader": "^0.5.8",
-    "webpack": "~2.6.1",
-    "webpack-dev-middleware": "^1.10.1",
-    "webpack-dev-server": "~2.4.2",
-    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
-    "webpack-merge": "~4.1.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/AngularClass/angular-starter.git"
-  },
-  "bugs": {
-    "url": "https://github.com/AngularClass/angular-starter/issues"
+    "@angular/cli": "1.3.2",
+    "@angular/compiler-cli": "4.3.6",
+    "@types/jasmine": "2.5.54",
+    "@types/node": "8.0.26",
+    "codelyzer": "3.1.2",
+    "jasmine-core": "2.8.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "1.7.1",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "1.0.1",
+    "karma-jasmine": "1.1.0",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "karma-coverage-istanbul-reporter": "1.3.0",
+    "protractor": "5.1.2",
+    "ts-node": "3.3.0",
+    "tslint": "5.7.0",
+    "typescript": "2.4.2"
   },
   "engines": {
-    "node": ">= 4.2.1",
-    "npm": ">= 3"
+    "node": ">= 6.9.0",
+    "npm": ">= 3.0.0"
   }
 }
